--- golang/HelloGo.ast
@@ -1,58 +1,0 @@
-     0  *ast.File {
-     1  .  Package: 2:1
-     2  .  Name: *ast.Ident {
-     3  .  .  NamePos: 2:9
-     4  .  .  Name: "main"
-     5  .  }
-     6  .  Decls: []ast.Decl (len = 1) {
-     7  .  .  0: *ast.FuncDecl {
-     8  .  .  .  Name: *ast.Ident {
-     9  .  .  .  .  NamePos: 4:6
-    10  .  .  .  .  Name: "main"
-    11  .  .  .  }
-    12  .  .  .  Type: *ast.FuncType {
-    13  .  .  .  .  Func: 4:1
-    14  .  .  .  .  Params: *ast.FieldList {
-    15  .  .  .  .  .  Opening: 4:10
-    16  .  .  .  .  .  Closing: 4:11
-    17  .  .  .  .  }
-    18  .  .  .  }
-    19  .  .  .  Body: *ast.BlockStmt {
-    20  .  .  .  .  Lbrace: 4:13
-    21  .  .  .  .  List: []ast.Stmt (len = 1) {
-    22  .  .  .  .  .  0: *ast.ExprStmt {
-    23  .  .  .  .  .  .  X: *ast.CallExpr {
-    24  .  .  .  .  .  .  .  Fun: *ast.Ident {
-    25  .  .  .  .  .  .  .  .  NamePos: 5:2
-    26  .  .  .  .  .  .  .  .  Name: "println"
-    27  .  .  .  .  .  .  .  }
-    28  .  .  .  .  .  .  .  Lparen: 5:9
-    29  .  .  .  .  .  .  .  Args: []ast.Expr (len = 1) {
-    30  .  .  .  .  .  .  .  .  0: *ast.BasicLit {
-    31  .  .  .  .  .  .  .  .  .  ValuePos: 5:10
-    32  .  .  .  .  .  .  .  .  .  Kind: STRING
-    33  .  .  .  .  .  .  .  .  .  Value: "\"Hello, Golang\\n\""
-    34  .  .  .  .  .  .  .  .  }
-    35  .  .  .  .  .  .  .  }
-    36  .  .  .  .  .  .  .  Ellipsis: -
-    37  .  .  .  .  .  .  .  Rparen: 5:27
-    38  .  .  .  .  .  .  }
-    39  .  .  .  .  .  }
-    40  .  .  .  .  }
-    41  .  .  .  .  Rbrace: 6:1
-    42  .  .  .  }
-    43  .  .  }
-    44  .  }
-    45  .  Scope: *ast.Scope {
-    46  .  .  Objects: map[string]*ast.Object (len = 1) {
-    47  .  .  .  "main": *ast.Object {
-    48  .  .  .  .  Kind: func
-    49  .  .  .  .  Name: "main"
-    50  .  .  .  .  Decl: *(obj @ 7)
-    51  .  .  .  }
-    52  .  .  }
-    53  .  }
-    54  .  Unresolved: []*ast.Ident (len = 1) {
-    55  .  .  0: *(obj @ 24)
-    56  .  }
-    57  }
