+++ golang/printast.go
@@ -5,40 +5,26 @@ import (
 	"go/parser"
 	"go/token"
 	"os"
-	"reflect"
 )
 
-func main() {
-	// src is the input for which we want to print the AST.
-	src := `
-package main
-
-
-
-func main() {
-	x := 1
-	println(x)
-
+func check(e error) {
+	if e != nil {
+		panic(e)
+	}
 }
-`
 
-	// Create the AST by parsing src.
-	fset := token.NewFileSet() // positions are relative to fset
-	f, err := parser.ParseFile(fset, "", src, 0)
-	if err != nil {
-		panic(err)
+func main() {
+	//TODO: add away to write the files to a text
+	files := os.Args[1:]
+	// Create the AST by go src.
+	fset := token.NewFileSet()
+	for i := 0; i < len(files); i++ {
+		f, err := parser.ParseFile(fset, string(files[i]), nil, parser.ParseComments)
+		check(err)
+		ast.Print(fset, f)
+		// d1 := []byte(asts)
+		// err := ioutil.WriteFile("/tmp/"+string(files[i]), d1, 0644)
+		// check(err)
 	}
 
-	// Print the AST.
-	// ast.Print(fset, f)
-	f2, err2 := os.Create("XequalsOne.ast")
-	if err2 != nil {
-		panic(err2)
-	}
-	ast.Fprint(f2, fset, f, func(name string, value reflect.Value) bool {
-		if ast.NotNilFilter(name, value) {
-			return value.Type().String() != "*ast.Object"
-		}
-		return false
-	})
 }
